Rename modules: message_router -> multithreading_router
